+++ src/game/config/characters.js
@@ -108,7 +108,10 @@ export const CHARACTERS = [
     sprite: 'char-guiri',
     stats: { peso: 4, equilibrio: 1, altura: 5, edad: 5 },
     available: true,
-    skins: [{ spritesheet: 'guiri', nombre: 'Clásico', flags: null }],
+    skins: [
+      { spritesheet: 'guiri', nombre: 'Clásico', flags: null },
+      { spritesheet: 'guiri_trampa', nombre: 'Make Triana Great Again', flags: 10 },
+    ],
   },
   {
     id: 'cunaos',
@@ -138,6 +141,19 @@ export const CHARACTERS = [
       { spritesheet: 'retro01', nombre: 'El Faraón', flags: 3 },
       { spritesheet: 'retro03', nombre: 'Dan', flags: 9 },
       { spritesheet: 'larry', nombre: 'El Detective', flags: 15 },
+    ],
+  },
+  {
+    id: 'curro',
+    name: 'CURRO',
+    description: 'Hola, soy Curro, la mascota de la Expo 92',
+    sprite: 'char-curro',
+    stats: { peso: 9, equilibrio: 2, altura: 9, edad: 9 },
+    available: true,
+    skins: [
+      { spritesheet: 'curro', nombre: 'Curro', flags: null },
+      { spritesheet: 'guasi_curro', nombre: 'Guasi Curro', flags: 5 },
+      { spritesheet: 'triana_curro', nombre: 'Triana Curro', flags: 10 },
     ],
   },
   {